% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboardShortcut.R
\name{keyboardShortcut}
\alias{keyboardShortcut}
\title{Add A Keyboard Shortcut}
\usage{
keyboardShortcut(inputId, key, ctrl = FALSE, alt = FALSE,
  shift = FALSE, description = NULL)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{key}{The key that must be pressed to trigger the keyboard shortcut}

\item{ctrl}{Does the Ctrl button have to be pressed?}

\item{alt}{Does the Alt button have to be pressed?}

\item{shift}{Does the shift button have to be pressed?}

\item{description}{A description of what the shortcut does to include in the legend}
}
\description{
Creates a keyboard shortcut in a shiny app. It binds like an actionButton or actionLink. So
you can trigger reactive events like you would with a button or link.
}
