% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExampleSKS.R
\name{runExampleSKS}
\alias{runExampleSKS}
\title{Run shinyKeyboardShortcuts examples}
\usage{
runExampleSKS(example = "old_faithful", port = NULL,
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"),
  display.mode = "showcase")
}
\arguments{
\item{example}{The example to run (currently only 'old_faithful')}

\item{port}{The TCP port that the application should listen on. Defaults to choosing a random port.}

\item{launch.browser}{If true, the system's default web browser will be launched automatically after the app is started. Defaults to true in interactive sessions only.}

\item{host}{The IPv4 address that the application should listen on. Defaults to the shiny.host option, if set, or "127.0.0.1" if not.}

\item{display.mode}{The mode in which to display the example. Defaults to showcase, but may be set to normal to see the example without code or commentary.}
}
\description{
Run an example that shows capabilities of shinyKeyboardShortcuts
}
