% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{shortcut_legend}
\alias{shortcut_legend}
\title{Create a Shortcut Legend Entry}
\usage{
shortcut_legend(..., animated = TRUE, description, size = NULL)
}
\arguments{
\item{\dots}{The keys that should be pressed to activate the shortcut}

\item{animated}{Should the legend be animated?}

\item{description}{What does the shortcut do?}

\item{size}{The size of the shortcut `lg` for large, `sm` for small, or missing for regular}
}
\description{
This function allows you to add a description of a keyboard shortcut to your app.
}
